Updated the slicer
- made the slicer into an external library class
- made the slicer tool depend upon the library
- this is since I would like to mesh from the polylines
